Implement numba fast dithers
